+++ relay-server/src/processing/mod.rs
@@ -9,7 +9,6 @@
 use relay_config::{Config, RelayMode};
 use relay_dynamic_config::GlobalConfig;
 use relay_quotas::RateLimits;
-use relay_sampling::evaluation::ReservoirCounters;
 
 use crate::managed::{Counted, Managed, ManagedEnvelope, Rejected};
 use crate::metrics_extraction::ExtractedMetrics;
@@ -87,11 +86,6 @@ pub struct Context<'a> {
     ///
     /// The caller needs to ensure the rate limits are not yet expired.
     pub rate_limits: &'a RateLimits,
-
-    /// Counters used for getting more samples for a project on-demand.
-    ///
-    /// Reservoir counters are a legacy feature and will be removed in the near future.
-    pub reservoir_counters: &'a ReservoirCounters,
 }
 
 impl<'a> Context<'a> {
@@ -134,7 +128,6 @@ impl Context<'static> {
         static GLOBAL_CONFIG: LazyLock<GlobalConfig> = LazyLock::new(Default::default);
         static PROJECT_INFO: LazyLock<ProjectInfo> = LazyLock::new(Default::default);
         static RATE_LIMITS: LazyLock<RateLimits> = LazyLock::new(Default::default);
-        static RESERVOIR_COUNTERS: LazyLock<ReservoirCounters> = LazyLock::new(Default::default);
 
         Self {
             config: &CONFIG,
@@ -142,7 +135,6 @@ impl Context<'static> {
             project_info: &PROJECT_INFO,
             sampling_project_info: None,
             rate_limits: &RATE_LIMITS,
-            reservoir_counters: &RESERVOIR_COUNTERS,
         }
     }
 }