+++ relay-server/src/services/processor.rs
@@ -27,7 +27,7 @@ use relay_event_schema::protocol::{ClientReport, EventId, SpanV2};
 use relay_filter::FilterStatKey;
 use relay_metrics::{Bucket, BucketMetadata, BucketView, BucketsView, MetricNamespace};
 use relay_quotas::{DataCategory, RateLimits, Scoping};
-use relay_sampling::evaluation::{ReservoirCounters, SamplingDecision};
+use relay_sampling::evaluation::SamplingDecision;
 use relay_statsd::metric;
 use relay_system::{Addr, FromMessage, NoResponse, Service};
 use reqwest::header;
@@ -674,8 +674,6 @@ pub struct ProcessEnvelope {
     pub rate_limits: Arc<RateLimits>,
     /// Root sampling project info.
     pub sampling_project_info: Option<Arc<ProjectInfo>>,
-    /// Sampling reservoir counters.
-    pub reservoir_counters: ReservoirCounters,
 }
 
 /// Like a [`ProcessEnvelope`], but with an envelope which has been grouped.
@@ -988,8 +986,8 @@ impl EnvelopeProcessorService {
         }
 
         #[cfg(feature = "processing")]
-        let rate_limiter = redis.as_ref().map(|redis| {
-            RedisRateLimiter::new(redis.quotas.clone())
+        let rate_limiter = redis.map(|redis| {
+            RedisRateLimiter::new(redis.quotas)
                 .max_limit(config.max_rate_limit())
                 .cache(config.quota_cache_ratio(), config.quota_cache_max())
         });
@@ -1026,10 +1024,6 @@ impl EnvelopeProcessorService {
                 transactions: TransactionProcessor::new(
                     Arc::clone(&quota_limiter),
                     geoip_lookup.clone(),
-                    #[cfg(feature = "processing")]
-                    redis.map(|r| r.quotas),
-                    #[cfg(not(feature = "processing"))]
-                    None,
                 ),
                 profile_chunks: ProfileChunksProcessor::new(Arc::clone(&quota_limiter)),
                 trace_attachments: TraceAttachmentsProcessor::new(Arc::clone(&quota_limiter)),
@@ -1360,7 +1354,6 @@ impl EnvelopeProcessorService {
                 project_info: &message.project_info,
                 sampling_project_info: message.sampling_project_info.as_deref(),
                 rate_limits: &message.rate_limits,
-                reservoir_counters: &message.reservoir_counters,
             };
 
             let message = ProcessEnvelopeGrouped {