+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Enable OTLP endpoints by default. ([#5951](https://github.com/getsentry/relay/pull/5951))
 - Backfill `app.vitals.start.screen` for V1 app-start transactions from the transaction name. ([#5960](https://github.com/getsentry/relay/pull/5960))
 - Enable performance score calculation for V2 spans. ([#5947](https://github.com/getsentry/relay/pull/5947))
+- Update user agent parsing rules. ([#5999](https://github.com/getsentry/relay/pull/5999))
 
 **Bug Fixes**:
 
@@ -19,6 +20,7 @@
 **Internal**:
 
 - Remove unknown debug image variants in errors. ([#5962](https://github.com/getsentry/relay/pull/5962))
+- Remove support for dynamic sampling reservoir rules. ([#5988](https://github.com/getsentry/relay/pull/5988))
 - Bump `sentry-conventions` to 0.6.0-4. ([#5944](https://github.com/getsentry/relay/pull/5944))
 - Bump `sqlparser` to 0.62. ([#5964](https://github.com/getsentry/relay/pull/5964))
 - Enable compression for forwarded uploads. ([#5965](https://github.com/getsentry/relay/pull/5965))